extract VideoRecorderToolbar as a component
